Change ValueError to TypeError for P5 * P8, etc.
TypeError makes more sense

I can't trigger AttributeError, maybe it's from an old Python version.

ChatGPT:

"There is no native Python object that would raise an AttributeError
when trying to cast it to an integer using int() or when trying to use
it as an argument in the __mul__ method of Interval. All native Python
objects that can be cast to an integer, such as integers, floating point
numbers, and strings representing numbers, will not raise an
AttributeError."